// src/game/domain/enums/CombatEventType.ts
export enum CombatEventType {
    PLAYER_HIT = 'PLAYER_HIT',
    PLAYER_PRECISE_HIT = 'PLAYER_PRECISE_HIT',
    PLAYER_MISS = 'PLAYER_MISS',
    PLAYER_TOOK_DAMAGE = 'PLAYER_TOOK_DAMAGE',
    PLAYER_EVADED = 'PLAYER_EVADED',
    PLAYER_EVADE_FAILED = 'PLAYER_EVADE_FAILED',
    PLAYER_COUNTER = 'PLAYER_COUNTER',
}